As title. Introduce fusion patterns to enable fusing quantized convolution and linear graph patterns into a custom op.

## Changes

Introduce the concept of using custom pattern detection functions to detect graph patterns rather than solely relying on SubgraphMatcher. The issue with SubgraphMatcher is that a large number of graph patterns may need to be exported to obtain variants for different combinations of decompositions/quantization workflows. Having a custom detection function improves maintainability.

Implement detection + replacement functions for quantized linear and quantized conv2d.

Differential Revision: [D81323425](https://our.internmc.facebook.com/intern/diff/D81323425/)

[ghstack-poisoned]

+++ backends/vulkan/patterns/quantized_linear.py
@@ -314,7 +314,7 @@ def make_linear_q8ta_q8csw_custom_op(
         weight_tensor_name = utils.get_tensor_name(ep, match.weight_node)
         # Pre-compute the weight sums which are needed to apply activation zero point
         # when using integer accumulation.
-        sum_per_output_channel = weight_tensor.sum(dim=1).to(torch.float).contiguous()
+        sum_per_output_channel = weight_tensor.sum(dim=1).to(torch.int32).contiguous()
         sums_name = weight_tensor_name + "_sums"
         # Sanitize the name
         sums_name = sums_name.replace(".", "_")
